Part.1 - Refactoring wriiten before code, organize statemanagement
