+++ src/state/location.js
@@ -1,0 +1,13 @@
+// import axios, * as others from 'axios'
+
+// export const getSityandCountry = async position => {
+//     let apiUrl = `https://geocode.xyz/${position.coords.latitude},${position.coords.longitude}?json=1`
+//     await axios
+//       .get(apiUrl)
+//       .then(Result => {
+//         locationSuccess(Result)
+//       })
+//       .catch(Error => {
+//         locationError()
+//       })
+//   }
